Corrected filter template loading.

+++ apps/opencs/model/doc/document.cpp
@@ -2245,29 +2245,33 @@ CSMDoc::Document::Document (const Files::ConfigurationManager& configuration, co
         mData.setDescription ("");
         mData.setAuthor ("");
     }
-/// \todo un-outcomment the else, once loading an existing content file works properly again.
+    if (boost::filesystem::exists (mProjectPath))
+    {
+        std::cout<<"Loading file."<<std::endl;
+        getData().loadFile (mProjectPath, false, true);
+    }
     else
     {
-        if (boost::filesystem::exists (mProjectPath))
+        std::cout<<"Attemp to use filters file as template."<<std::endl;
+        boost::filesystem::path locCustomFiltersPath (configuration.getUserPath());
+        locCustomFiltersPath /= "defaultfilters";
+        if (boost::filesystem::exists(locCustomFiltersPath))
         {
-            getData().loadFile (mProjectPath, false, true);
-        }
-        else
-        {
-            boost::filesystem::path locCustomFiltersPath (configuration.getUserPath());
-            locCustomFiltersPath /= "defaultfilters";
-            if (boost::filesystem::exists(locCustomFiltersPath))
+            std::cout<<"Found custom filters file ("<<locCustomFiltersPath<<")."<<std::endl;
+            boost::filesystem::copy_file (locCustomFiltersPath, mProjectPath);
+        } else {
+            boost::filesystem::path filters(mResDir);
+            filters /= "defaultfilters";
+            std::cout<<"Attempt to read defaultfilters from: "<<filters<<std::endl;
+            if (boost::filesystem::exists(filters))
             {
-                boost::filesystem::copy_file (locCustomFiltersPath, mProjectPath);
-            } else {
-                boost::filesystem::path filters(mResDir);
-                filters /= "defaultfilters";
                 boost::filesystem::copy_file(filters, mProjectPath);
+            } else {
+                std::cout<<filters<<" could not be found!"<<std::endl;
             }
             getData().loadFile (mProjectPath, false, true);
         }
     }
-
     addOptionalGmsts();
     addOptionalGlobals();
